requirements.txt
flask
flask-cors
pandas
openpyxl
pypdf
openai
python-dotenv
werkzeug
